extract: choose step behavior
Everything related to step selection,
initial step decision and navigation between
steps had been extracted into its file.
