+++ src/features/createAndEditTemplate/workflow/ui/AddOrEditTemplate.tsx
@@ -6,7 +6,6 @@ import {
   Wizard,
   WizardHeader,
   WizardStep,
-  useWizardContext,
 } from '@patternfly/react-core';
 
 import { useSearchParams } from 'react-router-dom';
@@ -20,8 +19,8 @@ import DetailStep from '../../describeTemplate/ui/DetailStep';
 import ReviewStep from '../../reviewTemplateRequest/ui/ReviewStep';
 import { isEmpty } from 'lodash';
 import { createUseStyles } from 'react-jss';
-import { useEffect } from 'react';
 import { useOnCancelModal } from '../core/cancelModal';
+import { useInitialStep, WizardUrlSync } from '../core/chooseStep';
 
 const useStyles = createUseStyles({
   minHeightForSpinner: {
@@ -33,39 +32,6 @@ export interface Props {
   isDisabled: boolean;
   addRepo: (snapshot: boolean) => void;
 }
-
-const DEFAULT_STEP_ID = 'define-content';
-
-const stepIdToIndex: Record<string, number> = {
-  'define-content': 2,
-  'redhat-repositories': 3,
-  'custom-repositories': 4,
-  'set-up-date': 5,
-  detail: 6,
-  review: 7,
-};
-
-// Component to sync URL with wizard state (must be inside Wizard)
-const WizardUrlSync = ({ onCancel }: { onCancel: () => void }) => {
-  const { goToStepById, activeStep } = useWizardContext();
-  const [urlSearchParams] = useSearchParams();
-
-  const tabParam = urlSearchParams.get('tab');
-
-  useEffect(() => {
-    if (tabParam && tabParam !== activeStep?.id) {
-      if (stepIdToIndex[tabParam]) {
-        goToStepById(tabParam);
-      } else if (tabParam === 'content') {
-        goToStepById(DEFAULT_STEP_ID);
-      } else {
-        onCancel();
-      }
-    }
-  }, [tabParam, activeStep?.id, goToStepById, onCancel]);
-
-  return null;
-};
 
 type TemplateBaseProps = {
   modalProps: {
@@ -86,24 +52,16 @@ type TemplateBaseProps = {
 
 const AddOrEditTemplateBase = ({ modalProps, wizardHeaderProps, footer }: TemplateBaseProps) => {
   const classes = useStyles();
-  const [urlSearchParams, setUrlSearchParams] = useSearchParams();
+  const [, setUrlSearchParams] = useSearchParams();
 
   const { isEdit, templateRequest, checkIfCurrentStepValid, editUUID } = useAddTemplateContext();
 
+  const initialIndex = useInitialStep();
   const onCancel = useOnCancelModal();
 
   // useSafeUUIDParam in AddTemplateContext already validates the UUID
   // If in edit mode and UUID is invalid, it will be an empty string
   if (isEdit && !editUUID) throw new Error('UUID is invalid');
-
-  const tabParam = urlSearchParams.get('tab');
-  const initialIndex = tabParam && stepIdToIndex[tabParam] ? stepIdToIndex[tabParam] : 2;
-
-  useEffect(() => {
-    if (!urlSearchParams.get('tab')) {
-      setUrlSearchParams({ tab: DEFAULT_STEP_ID }, { replace: true });
-    }
-  }, []);
 
   const sharedFooterProps = {
     nextButtonProps: { ouiaId: 'wizard-next-btn' },
